+++ src/Controller/JobController.php
@@ -29,6 +29,9 @@ class JobController extends Controller
     public function new(Request $request): Response
     {
         $job = new Job();
+
+        $this->denyAccessUnlessGranted('EDIT',$job);
+
         $form = $this->createForm(JobType::class, $job);
         $form->handleRequest($request);
 
@@ -51,6 +54,8 @@ class JobController extends Controller
      */
     public function show(Job $job): Response
     {
+        $this->denyAccessUnlessGranted('SHOW',$job);
+
         return $this->render('job/show.html.twig', ['job' => $job]);
     }
 
@@ -59,6 +64,8 @@ class JobController extends Controller
      */
     public function edit(Request $request, Job $job): Response
     {
+        $this->denyAccessUnlessGranted('EDIT',$job);
+
         $form = $this->createForm(JobType::class, $job);
         $form->handleRequest($request);
 
@@ -79,6 +86,8 @@ class JobController extends Controller
      */
     public function delete(Request $request, Job $job): Response
     {
+        $this->denyAccessUnlessGranted('DELETE',$job);
+
         if ($this->isCsrfTokenValid('delete'.$job->getId(), $request->request->get('_token'))) {
             $em = $this->getDoctrine()->getManager();
             $em->remove($job);
