+++ src/Controller/RoadSectionController.php
@@ -5,6 +5,8 @@ namespace App\Controller;
 use App\Entity\RoadSection;
 use App\Form\RoadSectionType;
 use App\Repository\RoadSectionRepository;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -52,6 +54,9 @@ class RoadSectionController extends Controller
     public function new(Request $request): Response
     {
         $roadSection = new RoadSection();
+
+        $this->denyAccessUnlessGranted('EDIT',$roadSection);
+
         $form = $this->createForm(RoadSectionType::class, $roadSection);
         $form->handleRequest($request);
 
@@ -74,6 +79,8 @@ class RoadSectionController extends Controller
      */
     public function show(RoadSection $roadSection): Response
     {
+        $this->denyAccessUnlessGranted('SHOW',$roadSection);
+
         return $this->render('road_section/show.html.twig', ['road_section' => $roadSection]);
     }
 
@@ -82,6 +89,8 @@ class RoadSectionController extends Controller
      */
     public function edit(Request $request, RoadSection $roadSection): Response
     {
+        $this->denyAccessUnlessGranted('EDIT',$roadSection);
+
         $form = $this->createForm(RoadSectionType::class, $roadSection);
         $form->handleRequest($request);
 
@@ -102,6 +111,8 @@ class RoadSectionController extends Controller
      */
     public function delete(Request $request, RoadSection $roadSection): Response
     {
+        $this->denyAccessUnlessGranted('DELETE',$roadSection);
+
         if ($this->isCsrfTokenValid('delete'.$roadSection->getId(), $request->request->get('_token'))) {
             $em = $this->getDoctrine()->getManager();
             $em->remove($roadSection);
